feat(test): add serviceProviderMacro + renaming
