+++ api/providers/test/src/index.ts
@@ -1,1 +1,2 @@
 export { serviceProviderMacro } from './serviceProviderMacro';
+export { handlerMacro } from './handlerMacro';
